requirements.txt
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
build==1.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
dateparser==1.2.1
fonttools==4.59.2
frozenlist==1.7.0
idna==3.10
kiwisolver==1.4.9
multidict==6.4.4
packaging==25.0
pandas==2.3.0
pillow==11.3.0
propcache==0.3.2
pycryptodome==3.23.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
